Consistent naming for config reader variables.

+++ L2J_Mobius_CT_2.4_Epilogue/java/org/l2jmobius/gameserver/model/olympiad/OlympiadGame.java
@@ -1435,7 +1435,7 @@ class OlympiadGameTask implements Runnable
 		SystemMessage sm;
 		// Waiting for teleport to arena
 		byte step = 60;
-		for (byte i = Config.ALT_OLY_WAIT_TIME; i > 0; i -= step)
+		for (int i = Config.ALT_OLY_WAIT_TIME; i > 0; i -= step)
 		{
 			sm = new SystemMessage(SystemMessageId.YOU_WILL_BE_MOVED_TO_THE_OLYMPIAD_STADIUM_IN_S1_SECOND_S);
 			sm.addInt(i);
